fix(workflows): extract run_credentials to lightweight module + importorskip integration tests (#9037)
